# backend/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - multipart/form-data
# format of the responses to the client (Accepts)
produces:
  - application/json
  - multipart/form-data
paths:
...
  /v1/login:  
    x-swagger-router-controller: controller
    post:
      operationId: v1_login
      parameters:
        - name: data
          in: body
          required: true
          schema:
              $ref: "#/definitions/login_definition"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /v1/asset/upload:
    x-swagger-router-controller: controller
    post:
      consumes:
        - multipart/form-data
      operationId: v1_upload
      parameters:
        - name: uploadedFile
          in: formData
          type: file
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  login_definition:
    required:
      - email
      - password
      - role
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
...